+++ src/CG.Blazor.Forms/Attributes/HTML/RenderInputSelectAttribute.cs
@@ -207,7 +207,9 @@ namespace CG.Blazor.Forms.Attributes
                     var label = string.IsNullOrEmpty(Label) ? prop.Name : Label;
 
                     // Split the options.
-                    var options = Options.Split(',');
+                    var options = Options.Split(',')
+                        .Select(x => x.Trim())
+                        .ToArray();
 
                     // Render the outer div.
                     index = builder.RenderUIElement(